requirements.txt
branca==0.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
folium==0.10.1
future==0.18.2
geographiclib==1.50
geopy==1.21.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
